requirements.txt
python-telegram-bot[webhooks]==20.3
requests
pandas
openpyxl